* Moved utils.py to test folder.
